add comment

+++ wrap/igl/lscm_parametrization.h
@@ -69,7 +69,7 @@ void OptimizeUV_LSCM( MeshType& m ,
     vcg::tri::MeshToMatrix< MeshType >::GetTriMeshData( m, F, V );
 
     // build fixed points data
-    size_t nFixed = 0;
+    int nFixed = 0;
     for (int i=0; i<(int)m.vert.size(); i++) {
         if (m.vert[i].Flags()&fixedMask) nFixed++;
     }
